Updating Dependencies & Fixed Deprecated
