--- app/src/main/java/com/niyaj/popos/features/components/FilterItem.kt
@@ -1,68 +1,0 @@
-package com.niyaj.popos.features.components
-
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.RadioButtonChecked
-import androidx.compose.material.icons.filled.RadioButtonUnchecked
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import com.niyaj.popos.features.common.ui.theme.SpaceMedium
-import com.niyaj.popos.features.common.ui.theme.SpaceSmall
-
-@Composable
-fun FilterItem(
-    modifier: Modifier = Modifier,
-    filterName: String,
-    color: Color,
-    itemSelected: Boolean = false,
-    onSelected: () -> Unit,
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceEvenly,
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .clip(RoundedCornerShape(4.dp))
-                .border(1.dp, color, RoundedCornerShape(4.dp))
-                .clickable {
-                    onSelected()
-                }
-                .padding(SpaceSmall),
-
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = filterName,
-                style = MaterialTheme.typography.body2,
-                fontWeight = if (itemSelected) FontWeight.SemiBold else FontWeight.Medium
-            )
-
-            Spacer(modifier = Modifier.width(SpaceMedium))
-            Icon(
-                imageVector = if (itemSelected) Icons.Default.RadioButtonChecked else Icons.Default.RadioButtonUnchecked,
-                contentDescription = null,
-                tint = color
-            )
-        }
-    }
-}
