--- app/src/main/java/com/niyaj/popos/features/components/NiaLoadingWheel.kt
@@ -1,152 +1,0 @@
-package com.niyaj.popos.features.components
-
-import androidx.compose.animation.animateColor
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.FastOutSlowInEasing
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.StartOffset
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.keyframes
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.rotate
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.niyaj.popos.features.common.ui.theme.PoposTheme
-import kotlinx.coroutines.launch
-
-@Composable
-fun NiaLoadingWheel(
-    contentDesc: String,
-    modifier: Modifier = Modifier,
-) {
-    val infiniteTransition = rememberInfiniteTransition()
-
-    // Specifies the float animation for slowly drawing out the lines on entering
-    val startValue = if (LocalInspectionMode.current) 0F else 1F
-    val floatAnimValues = (0 until NUM_OF_LINES).map { remember { Animatable(startValue) } }
-    LaunchedEffect(floatAnimValues) {
-        (0 until NUM_OF_LINES).map { index ->
-            launch {
-                floatAnimValues[index].animateTo(
-                    targetValue = 0F,
-                    animationSpec = tween(
-                        durationMillis = 100,
-                        easing = FastOutSlowInEasing,
-                        delayMillis = 40 * index,
-                    ),
-                )
-            }
-        }
-    }
-
-    // Specifies the rotation animation of the entire Canvas composable
-    val rotationAnim by infiniteTransition.animateFloat(
-        initialValue = 0F,
-        targetValue = 360F,
-        animationSpec = infiniteRepeatable(
-            animation = tween(durationMillis = ROTATION_TIME, easing = LinearEasing),
-        ),
-    )
-
-    // Specifies the color animation for the base-to-progress line color change
-    val baseLineColor = MaterialTheme.colors.primary
-    val progressLineColor = MaterialTheme.colors.primaryVariant
-    val colorAnimValues = (0 until NUM_OF_LINES).map { index ->
-        infiniteTransition.animateColor(
-            initialValue = baseLineColor,
-            targetValue = baseLineColor,
-            animationSpec = infiniteRepeatable(
-                animation = keyframes {
-                    durationMillis = ROTATION_TIME / 2
-                    progressLineColor at ROTATION_TIME / NUM_OF_LINES / 2 with LinearEasing
-                    baseLineColor at ROTATION_TIME / NUM_OF_LINES with LinearEasing
-                },
-                repeatMode = RepeatMode.Restart,
-                initialStartOffset = StartOffset(ROTATION_TIME / NUM_OF_LINES / 2 * index),
-            ),
-        )
-    }
-
-    // Draws out the LoadingWheel Canvas composable and sets the animations
-    Canvas(
-        modifier = modifier
-            .size(48.dp)
-            .padding(8.dp)
-            .graphicsLayer { rotationZ = rotationAnim }
-            .testTag(contentDesc),
-    ) {
-        repeat(NUM_OF_LINES) { index ->
-            rotate(degrees = index * 30f) {
-                drawLine(
-                    color = colorAnimValues[index].value,
-                    // Animates the initially drawn 1 pixel alpha from 0 to 1
-                    alpha = if (floatAnimValues[index].value < 1f) 1f else 0f,
-                    strokeWidth = 4F,
-                    cap = StrokeCap.Round,
-                    start = Offset(size.width / 2, size.height / 4),
-                    end = Offset(size.width / 2, floatAnimValues[index].value * size.height / 4),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun NiaOverlayLoadingWheel(
-    contentDesc: String,
-    modifier: Modifier = Modifier,
-) {
-    Surface(
-        shape = RoundedCornerShape(60.dp),
-        elevation = 8.dp,
-        color = MaterialTheme.colors.surface.copy(alpha = 0.83f),
-        modifier = modifier
-            .size(60.dp),
-    ) {
-        NiaLoadingWheel(
-            contentDesc = contentDesc,
-        )
-    }
-}
-
-@Preview
-@Composable
-fun NiaLoadingWheelPreview() {
-    PoposTheme {
-        Surface {
-            NiaLoadingWheel(contentDesc = "LoadingWheel")
-        }
-    }
-}
-
-@Preview
-@Composable
-fun NiaOverlayLoadingWheelPreview() {
-    PoposTheme {
-        Surface {
-            NiaOverlayLoadingWheel(contentDesc = "LoadingWheel")
-        }
-    }
-}
-
-private const val ROTATION_TIME = 12000
-private const val NUM_OF_LINES = 12
