+++ app/src/main/java/com/niyaj/popos/features/customer/presentation/components/CustomerCard.kt
@@ -19,12 +19,12 @@ import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.Rule
 import androidx.compose.material.icons.filled.Dns
 import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.KeyboardArrowDown
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material.icons.filled.PhoneAndroid
-import androidx.compose.material.icons.filled.Rule
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -145,7 +145,7 @@ fun ImportExportCustomerBody(
                     onClick = onClickSelectAll
                 ) {
                     Icon(
-                        imageVector = Icons.Default.Rule,
+                        imageVector = Icons.AutoMirrored.Filled.Rule,
                         contentDescription = "Select All Customers"
                     )
                 }
