+++ app/src/main/java/com/niyaj/popos/features/product/presentation/components/ProductCard.kt
@@ -27,9 +27,9 @@ import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.Rule
 import androidx.compose.material.icons.filled.Dns
 import androidx.compose.material.icons.filled.KeyboardArrowDown
-import androidx.compose.material.icons.filled.Rule
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -175,7 +175,7 @@ fun ProductBody(
                 IconButton(
                     onClick = onClickSelectAll
                 ) {
-                    Icon(imageVector = Icons.Default.Rule, contentDescription = "Select All Product")
+                    Icon(imageVector = Icons.AutoMirrored.Filled.Rule, contentDescription = "Select All Product")
                 }
             },
             expand = {  modifier: Modifier ->
