+++ gradle/libs.versions.toml
@@ -1,12 +1,12 @@
 [versions]
-androidGradlePlugin = "8.2.0"
-kotlin = '1.9.21'
+androidGradlePlugin = "8.2.2"
+kotlin = '1.9.22'
 dagger-hilt-version = "2.50"
 hilt-version = '1.1.0'
-realm-version = "1.11.0"
+realm-version = "1.13.0"
 sentry-version = '3.14.0'
 sentry-android-version = '6.33.0'
-ksp-version = '1.9.21-1.0.16'
+ksp-version = '1.9.22-1.0.17'
 appsweep = 'latest.release'
 androidx-core-version = "1.12.0"
 junit = '4.13.2'
@@ -27,7 +27,7 @@ moshi-version = '1.15.0'
 truth-version = '1.1.5'
 mockk-version = "1.13.5"
 profileinstaller-version = '1.3.1'
-androidx-livedata-version = "1.5.4"
+androidx-livedata-version = "1.6.0"
 androidx-navigation-compose-version = "2.7.6"
 androidx-test-core-version = '1.5.0'
 androidx-test-runner = '1.5.2'
@@ -37,23 +37,23 @@ androidx-splash-version = "1.0.1"
 androidx-paging-version = "3.1.1"
 # Compose
 compose-version = '1.4.0-beta01'
-compose-bom = "2023.10.01"
+compose-bom = "2024.01.00"
 activity-compose = "1.8.2"
-lifecycle-version = "2.6.2"
-ui-test-junit = '1.5.4'
-kotlinComposeExtension = '1.4.1' # As per Kotlin version
+lifecycle-version = "2.7.0"
+ui-test-junit = '1.6.0'
+androidxComposeCompiler = '1.5.8' # As per Kotlin version
 # Benchmark Variables
 uiautomator = '2.2.0'
-benchmark-macro = "1.2.2"
-androidx-baselineprofile = "1.2.2"
-androidx-benchmark = "1.2.2"
-benchmark-junit4 = "1.2.2"
+benchmark-macro = "1.2.3"
+androidx-baselineprofile = "1.2.3"
+androidx-benchmark = "1.2.3"
+benchmark-junit4 = "1.2.3"
 
 play-integrity = '1.3.0'
 play-update = "2.1.0"
 coil-version = "2.5.0"
 play-scanner = "16.1.0"
-play-service = "18.2.0"
+play-service = "18.3.0"
 
 # AndroidX camera
 camera-version = "1.3.1"
@@ -149,6 +149,7 @@ pos-printer = { group = "com.github.DantSu", name = "ESCPOS-ThermalPrinter-Andro
 
 # Realm
 realm-library-base = { group = "io.realm.kotlin", name = "library-base", version.ref = "realm-version" }
+realm-library-sync = { group = "io.realm.kotlin", name = "library-sync", version.ref = "realm-version" }
 
 
 # debugImplementation because LeakCanary should only run in debug builds.
